% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMTLd_test.R
\name{RMTLd_test}
\alias{RMTLd_test}
\title{Comparing restricted mean survival time}
\usage{
RMTLd_test(time, status, group, alpha = 0.05, tau = NULL)
}
\arguments{
\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 1 = event of interest, 2 = competing event and 0 = right censored.}

\item{group}{The group indicator for comparison. The elements of this vector take either 1 or 0. Normally, 0 = control group, 1 = active treatment group.}

\item{alpha}{The default is 0.05. (1-\code{alpha}) confidence intervals are reported.}

\item{tau}{A scaler value to specify the truncation time point for the RMST calculation.
The defaultis the minimum of the maximum follow-up time of two groups}
}
\value{
an object of class RMTLd_test.

\item{tau}{the truncation time used in the analyses}

\item{Note}{a note regarding the truncation time}

\item{Base}{RMTL results in two groups, including RMTL and corresponding variance in group 0 and 1}

\item{Estimation}{RMSTd results between groups, including RMTLd and corresponding (1-\code{alpha}) confidence intervals.}

\item{Test}{Results of the log-rank test, Gray test and RMTLd test analyses.}
}
\description{
Performs two-sample comparisons using the restricted mean time lost(RMTL) as a summary measure of the cumulative incidence function under competing risks
}
\examples{
library(crRMTL)
data(Simdata)
# The time point is set as the minimum of the maximum follow-up time of two groups when tau = NULL.
RMTLd_test(Simdata$time, Simdata$status, Simdata$group, alpha = 0.05, tau = NULL)
# The time point is set as tau = 3.5.
RMTLd_test(Simdata$time, Simdata$status, Simdata$group, alpha = 0.05, tau = 3.5)

}
